upload and lock at same time
